Avoid vfork() bear traps We can't call setIOManagerPipe() in the vfork() child, because the change will be reflected in the parent too.  Add a large warning to that effect.
I tried changing vfork() to fork(), but it seems that this leads to a
different problem: the fork() sometimes takes so long that it gets
repeatedly interrupted by the timer signal and never makes progress.
I could disable the timer for a while, but decided to back off and fix
the vfork() version instead.