All new sync process functions now terminate on an exception (#2233)
Now all the functions that call a process synchronously have the same
behaviour. Previously just readProcess, readProcessWithExitCode did
this, now callProcess and callCommand do too.

If a thread running one of these functions gets an exception, including
async exceptions (such as from timeout or killThread), then the
external process gets terminated.

Introduce a helper function to implement this behaviour. Currently it
is not exposed to users, but that could be changed easily.

Authored-by: Duncan Coutts <duncan@well-typed.com>
Signed-off-by: Herbert Valerio Riedel <hvr@gnu.org>