Add bounds on base in test-suite stanza
GHC's build system runs `ghc-cabal check` on the cabal files for all of its
core libraries. One of the checks that this performs is to verify that base
has bounds. IMHO it's a bit silly that this is necessary even when base
is otherwise constrained by the library stanza so I've left the bounds
quite loose.