Give a usable buffer to _pipe on Windows
Fixes #181, the documentation seems to have been misread, the argument to `_pipe` is the number of bytes to reserve for pipe communications, not the number of pipes to create.  This ended up making pipes that can only send 2 bytes at a time.

Instead use 8k which is the size of the internal Buffers in `Base` which back `I/O` calls.