posix: Ensure that find_executable accounts for working directory
When `execvpe` is not available we use `find_executable` to search for
the executable prior to `fork`ing (fixing GHC #19994). However,
previously we failed to account for the fact that `exec` will be called
from a different working directory than that of the spawning process if
the user has set `cwd`.

Fix this by teaching `find_executable` to search relative to the
requested working directory.

Fixes #219.