configure: Ensure that posix_spawnp reports ENOENT correctly
It turns out that POSIX allows implementations to report
errors "lazily" by exiting the spawned process with code 127.

In particular, OpenBSD's implementation returns successfully if the
requested executable doesn't exist. This severely hampers our ability to
report errors accurately.

In this patch we add a `configure` test for sensible error-reporting
from `posix_spawnp` and disable it if the test fails.

Fixes #224.