testsuite/process011: Don't rely on Python
The refactor of `process011` performed in 8cd7e04, where
the test is taught to use `python3` instead of `sh`, has broken the test
on Debian 10. Specifically, for reasons that I don't yet understand,
`python3` exits with code 1 not code 2 when killing itself with
signal 2. Strangely, my tests on other distributions (e.g. NixOS)
suggest that Python's behavior in this case is rather inconsistent.

To avoid this, we now rather use a dedicated C program instead of
Python.

Fixes #241.