testsuite: Don't attempt to read from inherited fd in processT251
It turns out that this test is subtly broken. In particular, the
test will fail if any file is opened in the subprocess before the child
is run since the closed fd 0 may be reused for the new file.
This tends to happen in the threaded RTS due to the event manager's
control pipe (see GHC #22395). Unfortunately, it's not really clear how
else #251 can reliably be tested.