cbits: Ensure that fork pipe doesn't shadow std fds
Previously we would assume that the pipe used to communicate
errors from the child back to the parent did not shadow the standard
descriptors (stdin, stdout, and stderr). Somewhat surprisingly, this
appears to hold most platforms. However, OpenBSD appears to be a notable
exception. This lead to the failure of the `processT251` test.

Avoid relying on this assumption by `dup`ing the pipe fds until they end
up out of the standard fd range.

Closes #266.