cbits/fork_exec: Eliminate potential leakage of comm. pipes
Previously we used `fork` followed by `fcntl` to create a pipe with
`O_CLOEXEC`. However, this is potentially racy as another thread may
`fork` in between in these two calls, resulting in the pipe of the
former thread being available to the child of the latter.

Fix this by using `O_CLOEXEC` when creating the communications pipe.

Closes #287.