Move tests to separate process-tests package
This commit moves the process testsuite to a separate package.
The rationale is that later commits require the test-suite to use a
Custom setup in order to work around Cabal bug #9854. This introduces a
setup-depends dependency on the Cabal library, and we would very much
like to avoid the dependency cycle:

  Cabal depends on process
  process depends on Cabal

Instead, splitting up the test-suite, we have:

  Cabal depends on process
  process-tests depends on process and on Cabal

To run the test-suite, you can use either of the following commands:

> cabal run process-tests:test
> stack test process-tests

There are a few other auxiliary changes, such as:

  - Using Cabal version >= 2.4 in the .cabal files,
    and fixing the associated warnings,
  - Using the CPP mingw32_HOST_OS define to check for Windows;
    this avoids having to define the same variable twice in two
    different packages.