Windows: Implement mitigation for % expansion
Another aspect of the BatBadBut vulnerability that the previous patch
did not address is the fact that `%` cannot be easily escaped in a
double-quoted string. The vulnerability discussion suggests
transliterating `%` as `%%cd:~,%`, which should evaluate to an empty
string assuming that the `DelayedExpansion` registry setting is
disabled.

Here I have implemented this additional mitigation.